Move `safeHandler` to common.ts
