+++ services/server/src/server/apiv1/testartifacts/testartifacts.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { findLatestChainTest } from "./testartifacts.handlers";
-import { safeHandler } from "../controllers.common";
+import { safeHandler } from "../../common";
 
 const router: Router = Router();
