+++ services/server/src/server/apiv1/verification/etherscan/stateless/etherscan.stateless.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import { verifyFromEtherscan } from "./etherscan.stateless.handlers";
-import { checkPerfectMatch, safeHandler } from "../../../controllers.common";
-
+import { checkPerfectMatch } from "../../../controllers.common";
+import { safeHandler } from "../../../../common";
 const router: Router = Router();
 
 router
