+++ services/server/src/server/apiv1/verification/session-state/session-state.routes.ts
@@ -5,7 +5,7 @@ import {
   restartSessionEndpoint,
   addInputContractEndpoint,
 } from "./session-state.handlers";
-import { safeHandler } from "../../controllers.common";
+import { safeHandler } from "../../../common";
 
 const router: Router = Router();
