+++ services/server/src/server/apiv1/verification/solc-json/session/solc-json.session.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { addInputSolcJsonEndpoint } from "./solc-json.session.handlers";
-import { safeHandler } from "../../../controllers.common";
+import { safeHandler } from "../../../../common";
 
 const router: Router = Router();
