+++ services/server/src/server/apiv1/verification/solc-json/stateless/solc-json.stateless.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import { verifySolcJsonEndpoint } from "./solc-json.stateless.handlers";
-import { checkPerfectMatch, safeHandler } from "../../../controllers.common";
+import { checkPerfectMatch } from "../../../controllers.common";
+import { safeHandler } from "../../../../common";
 
 const router: Router = Router();
